Don't fetch new page when same URL with hash
