+++ src/Components/Web.JS/src/Services/NavigationEnhancement.ts
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 import { synchronizeDomContent } from '../Rendering/DomMerging/DomSync';
-import { attachProgrammaticEnhancedNavigationHandler, handleClickForNavigationInterception, hasInteractiveRouter, notifyEnhancedNavigationListners } from './NavigationUtils';
+import { attachProgrammaticEnhancedNavigationHandler, handleClickForNavigationInterception, hasInteractiveRouter, isSamePageWithHash, notifyEnhancedNavigationListners, performScrollToElementOnTheSamePage } from './NavigationUtils';
 
 /*
 In effect, we have two separate client-side navigation mechanisms:
@@ -89,8 +89,14 @@ function onDocumentClick(event: MouseEvent) {
   }
 
   handleClickForNavigationInterception(event, absoluteInternalHref => {
+    const shouldScrollToHash = isSamePageWithHash(absoluteInternalHref);
     history.pushState(null, /* ignored title */ '', absoluteInternalHref);
-    performEnhancedPageLoad(absoluteInternalHref, /* interceptedLink */ true);
+
+    if (shouldScrollToHash) {
+      performScrollToElementOnTheSamePage(absoluteInternalHref);
+    } else {
+      performEnhancedPageLoad(absoluteInternalHref, /* interceptedLink */ true);
+    }
   });
 }
 
@@ -132,10 +138,10 @@ function onDocumentSubmit(event: SubmitEvent) {
 
     event.preventDefault();
 
-    const url = new URL(event.submitter?.getAttribute('formaction') || formElem.action, document.baseURI); 
-    const fetchOptions: RequestInit = { method: method}; 
+    const url = new URL(event.submitter?.getAttribute('formaction') || formElem.action, document.baseURI);
+    const fetchOptions: RequestInit = { method: method};
     const formData = new FormData(formElem);
-   
+
     const submitterName = event.submitter?.getAttribute('name');
     const submitterValue = event.submitter!.getAttribute('value');
     if (submitterName && submitterValue) {
@@ -153,8 +159,8 @@ function onDocumentSubmit(event: SubmitEvent) {
     } else {
       // Setting request body and content-type header depending on enctype
       const enctype = event.submitter?.getAttribute('formenctype') || formElem.enctype;
-      if (enctype === 'multipart/form-data') { 
-        // Content-Type header will be set to 'multipart/form-data' 
+      if (enctype === 'multipart/form-data') {
+        // Content-Type header will be set to 'multipart/form-data'
         fetchOptions.body = formData;
       } else {
         fetchOptions.body = urlSearchParams;
